Separating Consent Summary and Popup
